requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# django-cryptography==1.1

psycopg2==2.9.9

pillow==10.2.0
requests==2.31.0

redis==5.0.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
celery==5.3.6
flower==2.0.1

drf-yasg==1.21.7

django-prometheus==2.3.1
django-cors-headers==4.3.1
django-redis==5.4.0

black==23.12.1
coverage==7.4.0

aiohttp==3.9.3

# eth-keys==0.5.0
# eth-account==0.12.0
# solana==0.33.0
# base58==2.1.1